feat:share to other user,update comment of prgress change

+++ app/controllers/todos_controller.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
 class TodosController < ApplicationController
-  before_action :find_todo_by_user, only: %i[destroy show activate up down]
+  before_action :find_todo_by_user, only: %i[destroy show activate up down share_todo update_progress]
+  before_action :find_share_by_user, only: %i[ up down]
   before_action :current_active, only: %i[activate down up destroy]
   before_action :list_todos, only: %i[create activate down up destroy]
+  before_action :current_user_shares, only: %i[ down up ]
   respond_to :html, :js
 
   # fetching all todos of current user
@@ -13,7 +15,14 @@ class TodosController < ApplicationController
     list_todos
     page_rendering
   end
-
+  #sharing to to other users
+  def share_todo
+    users= params[:users]
+     users.each do |i|
+       p i.to_i
+       User.find(i.to_i).todos << @todo
+     end
+  end
   # creating new todo object
   def new
     @todo = @current_user.todos.build
@@ -39,31 +48,52 @@ class TodosController < ApplicationController
 
   # changing priority to down
   def down
-    symbol = "up"
-    @todo.priority_switch(symbol,@todos)
+    symbol = "down"
+    @share.priority_switch(symbol,@shares)
     page_rendering
   end
 
   # changing priority to up
   def up
-    symbol = "down"
-    @up=@todo.priority_switch(symbol,@todos)
+    symbol = "up"
+    @up=@share.priority_switch(symbol,@shares)
+    @up = @up.todo
     page_rendering
   end
 
   # deleting selected todo
   def destroy
-    if @todo.destroy
-      page_rendering
+    shared = @todo.shares.find_by(user_id: @current_user.id)
+    if shared.is_owner
+      if @todo.destroy
+        page_rendering
+      else
+        flash.now[:danger] = 'Todo cannot be Deleted'
+        page_rendering
+      end
     else
-      flash.now[:danger] = 'Todo cannot be Deleted'
-      page_rendering
+      if shared.destroy
+        page_rendering
+      else
+        flash.now[:danger] = 'Share cannot be removed'
+        page_rendering
+      end
     end
   end
   def show
 
   end
-
+  def update_progress
+    p params
+    @comment = @todo.comments.build(description: params[:progress] ,user_id: @current_user.id)
+    if @comment.save!
+      flash[:success] = 'Comment inserted Successfully'
+      redirect_to todo_path(@todo)
+    else
+      flash[:warning] = 'Cannot insert a blank comment'
+      redirect_to todo_path(@todo)
+    end
+  end
   private
 
   # finding the todo which is in the current users ownership
@@ -71,11 +101,14 @@ class TodosController < ApplicationController
     @todo = @current_user.todos.find(params[:id])
   end
 
+  def find_share_by_user
+    @share = @current_user.shares.find_by(todo_id: params[:id])
+  end
   # finding all todo
   def list_todos
     @active = @active.nil? ? true : @active
-    @todos = @current_user.todos
-    .select_by_active(@active).sort_by_priority.search(@search)
+    @todos = @current_user.todos.search(@search)
+    .select_by_active(@active).sort_by_priority
     @todos = @todos.paginate(:page => params[:page], :per_page => 5 )
   end
 
@@ -94,4 +127,9 @@ class TodosController < ApplicationController
   def current_active
     @active = @todo.active
   end
+
+  def current_user_shares
+    @shares = @current_user.shares.joins("left outer join todos on todos.id=shares.todo_id and todos.active=#{@active}").sort_by_priority
+  end
+
 end
